+++ traffic_ops/traffic_ops_golang/servicecategory/servicecategories.go
@@ -40,21 +40,7 @@ type TOServiceCategory struct {
 }
 
 func (v *TOServiceCategory) GetLastUpdated() (*time.Time, bool, error) {
-	found := false
-	lastUpdated := time.Time{}
-	rows, err := v.APIInfo().Tx.Query(`select last_updated from service_category where name=$1`, v.Name)
-	if err != nil {
-		return nil, found, errors.New("querying last_updated: " + err.Error())
-	}
-	defer rows.Close()
-	if !rows.Next() {
-		return nil, found, errors.New("no resource found with this id")
-	}
-	found = true
-	if err := rows.Scan(&lastUpdated); err != nil {
-		return nil, found, errors.New("scanning last_updated: " + err.Error())
-	}
-	return &lastUpdated, found, nil
+	return api.GetLastUpdatedByName(v.APIInfo().Tx, v.Name, "service_category")
 }
 
 func (v *TOServiceCategory) SetLastUpdated(t tc.TimeNoMod) { v.LastUpdated = t }
@@ -149,12 +135,16 @@ func Update(w http.ResponseWriter, r *http.Request) {
 	}
 
 	var origSC TOServiceCategory
-	if err := inf.Tx.QueryRow(`SELECT name FROM service_category WHERE name = $1`, name).Scan(&origSC.Name); err != nil {
+	if err := inf.Tx.QueryRow(`SELECT name, last_updated FROM service_category WHERE name = $1`, name).Scan(&origSC.Name, &origSC.LastUpdated); err != nil {
 		if err == sql.ErrNoRows {
 			api.HandleErr(w, r, inf.Tx.Tx, http.StatusBadRequest, errors.New("no service category found with name "+name), nil)
 			return
 		}
 		api.HandleErr(w, r, inf.Tx.Tx, http.StatusInternalServerError, nil, err)
+		return
+	}
+	if !api.IsUnmodified(r.Header, origSC.LastUpdated.Time) {
+		api.HandleErr(w, r, inf.Tx.Tx, http.StatusPreconditionFailed, errors.New("service category could not be modified because the precondition failed"), nil)
 		return
 	}